finalizing user profile.
