+++ OptimizelySDK/Optimizely.cs
@@ -150,9 +150,9 @@ namespace OptimizelySDK
                 return null;
             }
             //DecisionService.GetVariation(experiment, userId, userAttributes);
-            if (!ValidatePreconditions(experiment, userId, userAttributes))
+            if (!DecisionService.IsValid(experiment, userId, userAttributes))
             {
-                Logger.Log(LogLevel.INFO, string.Format("Not activating user {0}.", userId));
+                //Logger.Log(LogLevel.INFO, string.Format("Not activating user {0}.", userId));
                 return null;
             }
