+++ OptimizelySDK/UserProfileService.cs
@@ -20,61 +20,6 @@ using System.Collections.Generic;
 namespace OptimizelySDK
 {
 
-    public class UserProfile
-    {
-        /// <summary>
-        /// A user's ID.
-        /// </summary>
-        public string UserId;
-        /// <summary>
-        /// The bucketing experimentBucketMap of the user.
-        /// </summary>
-        public Dictionary<string, Decision> ExperimentBucketMap;
-
-        /// <summary>
-        /// Construct a User Profile instance from explicit components.
-        /// </summary>
-        /// <param name="userId">The ID of the user.</param>
-        /// <param name="experimentBucketMap">The bucketing experimentBucketMap of the user.</param>
-        public UserProfile(string userId, Dictionary<string, Decision> experimentBucketMap)
-        {
-            this.UserId = userId;
-            this.ExperimentBucketMap = experimentBucketMap;
-        }
-
-        public int HashCode()
-            {
-                int result = UserId.GetHashCode();
-                result = 31 * result + ExperimentBucketMap.GetHashCode();
-                return result;
-        }
-
-
-        /// <summary>
-        /// Convert a User Profile instance to a Map.
-        /// </summary>
-        /// <returns>A map representation of the user profile instance.</returns>
-        public Dictionary<string, object> ToMap()
-        {
-            Dictionary<string, object> userProfileMap = new Dictionary<string, object>();
-
-            userProfileMap[UserProfileService.USER_ID_KEY] = UserId;
-
-            var decisionsMap = new Dictionary<string, Dictionary<string, string>>();
-            //var decisionMap = new Dictionary<string, object>();
-            
-            foreach(var row in ExperimentBucketMap)
-            {
-                decisionsMap[row.Key] = row.Value.ToMap();
-            }
-
-            userProfileMap[UserProfileService.EXPERIMENT_BUCKET_MAP_KEY] = decisionsMap;
-
-            return userProfileMap;
-        }
-    }
-
-
     /// <summary>
     /// Class encapsulating user profile service functionality.
     /// Override with your own implementation for storing and retrieving the user profile.
@@ -93,78 +38,4 @@ namespace OptimizelySDK
 
         public abstract void Save(Dictionary<string, object> userProfile);
     }
-
-    public class UserProfileUtils
-    {
-        
-        /// <summary>
-        /// Validate whether a {@code Map<String, Object>} can be transformed into a {@link UserProfile}.
-        /// </summary>
-        /// <param name="map"The map to check.></param>
-        /// <returns>True if the map can be converted into a {@link UserProfile}. False if the map cannot be converted.</returns>
-        public static bool IsValidUserProfileMap(Dictionary<string, object> map)
-        {
-            // The Map must contain a value for the user ID
-            if (!map.ContainsKey(UserProfileService.USER_ID_KEY))
-            {
-                return false;
-            }
-            // The Map must contain a value for the experiment bucket map
-            if (!map.ContainsKey(UserProfileService.EXPERIMENT_BUCKET_MAP_KEY))
-            {
-                return false;
-            }
-            // The value for the experimentBucketMapKey must be a map
-            //if (!(map[UserProfileService.EXPERIMENT_BUCKET_MAP_KEY] is Dictionary<string, object>))
-            //{
-            //    return false;
-            //}
-            // Try and cast the experimentBucketMap value to a typed map
-            Dictionary<string, Dictionary<string, string>> experimentBucketMap;
-            try
-            {
-                experimentBucketMap = (Dictionary<string, Dictionary<string, string>>)map[UserProfileService.EXPERIMENT_BUCKET_MAP_KEY];
-                if (experimentBucketMap.Values.Count == 0) return false;
-            }
-            catch (Exception e)
-            {
-                return false;
-            }
-
-            foreach (Dictionary<string, string> decision in experimentBucketMap.Values)
-            {
-                if (!decision.ContainsKey(UserProfileService.VARIATION_ID_KEY))
-                {
-                    return false;
-                }
-            }
-
-            // the map is good enough for us to use
-            return true;
-        }
-
-        
-
-        /// <summary>
-        /// Convert a Map to a {@link UserProfile} instance.
-        /// </summary>
-        /// <param name="map">The map to construct the { @link UserProfile }</param>
-        /// <returns>A {@link UserProfile}instance.</returns>
-        public static UserProfile ConvertMapToUserProfile(Dictionary<string, object> map)
-        {
-            string userId = (string)map[UserProfileService.USER_ID_KEY];
-            
-            var experimentBucketMap = (Dictionary<string, Dictionary<string, string>>)map[UserProfileService.EXPERIMENT_BUCKET_MAP_KEY];
-
-            Dictionary<string, Decision> decisions = new Dictionary<string, Decision>();
-
-            foreach(var entry in experimentBucketMap)
-            {
-                Decision decision = new Decision(entry.Value[UserProfileService.VARIATION_ID_KEY]);
-                decisions[entry.Key] = decision;
-            }
-            return new UserProfile(userId, decisions);
-        }
-    }
-
 }
